backend/requirements.txt
uvicorn==0.34.3
gunicorn==23.0.0
fastapi==0.115.14
requests==2.32.4
mongoengine==0.29.1
PyJWT==2.9.0
pydantic-settings==2.10.1
pydantic[email]==2.10.1
passlib[bcrypt]==1.7.4
